move test_events2 & test_events3 into test_events
we dont need 3 files, one is enough
